% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_apparent_temperature.R
\name{cal_apparent_temperature}
\alias{cal_apparent_temperature}
\title{Calculate apparent temperature}
\usage{
cal_apparent_temperature(t, td, t_unit)
}
\arguments{
\item{t, td}{Numeric. Air temperature and dewpoint temperature in degree Celsius (°C),
degree Fahrenheit (°F), or Kelvin (K).}

\item{t_unit}{Character string. Unit of the temperature values.
Must be one of \code{"celsius"}, \code{"fahrenheit"}, or \code{"kelvin"}.}
}
\value{
Numeric value of apparent temperature in degree Celsius (°C).
}
\description{
Calculate apparent temperature \eqn{AT} (°C) from air temperature
\eqn{T} (°C) and dewpoint temperature \eqn{T_d} (°C).
}
\details{
Apparent temperature \eqn{AT} (°C) is an index used to assess
human comfort in the warm season.
It is calculated using the formula:

\deqn{AT = -2.653 + 0.994 \cdot T + 0.0153 \cdot T_d^2}

where \eqn{T} is the air temperature and \eqn{T_d} is the dewpoint temperature,
in degrees Celsius (°C).

The temperature input is converted to degree Celsius (°C) based on
the provided \code{t_unit} before calculation.
}
\examples{
cal_apparent_temperature(t = 28, td = 12, t_unit = "celsius")
cal_apparent_temperature(t = 82.4, td = 53.6, t_unit = "fahrenheit")

}
\references{
Davis, Robert E., Glenn R. McGregor, and Kyle B. Enfield.
“Humidity: A Review and Primer on Atmospheric Moisture and Human Health.”
Environmental Research 144 (January 2016): 106–16.
\url{https://doi.org/10.1016/j.envres.2015.10.014}.

Kalkstein, Laurence S., and Kathleen M. Valimont.
“An Evaluation of Summer Discomfort in the United State Using a Relative Climatological Index.”
Bulletin of the American Meteorological Society 67, no. 7 (July 1986): 842–48.
\url{https://doi.org/10.1175/1520-0477(1986)067<0842:AEOSDI>2.0.CO;2}.
}
\seealso{
\code{\link[=cal_dewpoint_temperature]{cal_dewpoint_temperature()}}
}
