% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_mixing_ratio.R
\name{cal_mixing_ratio}
\alias{cal_mixing_ratio}
\title{Calculate mixing ratio}
\usage{
cal_mixing_ratio(e, p)
}
\arguments{
\item{e}{Numeric. Vapor pressure in hectoPascal (hPa).}

\item{p}{Numeric. Air pressure in hectoPascal (hPa).}
}
\value{
Numeric value of mixing ratio in g/kg.
}
\description{
Calculate mixing ratio \eqn{w} (g/kg) from
vapor pressure \eqn{e} (hPa) and air pressure \eqn{p} (hPa).
}
\details{
Mixing ratio \eqn{w} (g/kg) is defined as the mass of moisture per mass of
dry air:

\deqn{q = \frac{0.622 \cdot e \cdot 1000}{p - e}}

where \eqn{e} is the vapor pressure (hPa), and \eqn{p} is the air pressure (hPa).
}
\examples{
cal_mixing_ratio(e = 12, p = 1013.25)

}
\references{
Davis, Robert E., Glenn R. McGregor, and Kyle B. Enfield.
“Humidity: A Review and Primer on Atmospheric Moisture and Human Health.”
Environmental Research 144 (January 2016): 106–16.
\url{https://doi.org/10.1016/j.envres.2015.10.014}.
}
