% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_relative_humidity.R
\name{cal_relative_humidity}
\alias{cal_relative_humidity}
\title{Calculate relative humidity}
\usage{
cal_relative_humidity(e, es)
}
\arguments{
\item{e}{Numeric. Actual humidity value. Can be vapor pressure, mixing ratio,
specific humidity, or absolute humidity.}

\item{es}{Numeric. Saturation value corresponding to \code{e},
using the same unit and type.}
}
\value{
Numeric value of relative humidity (\%).
}
\description{
Calculate relative humidity \eqn{RH} (\%) from actual and saturation values of
vapor pressure, mixing ratio, specific humidity, or absolute humidity.
}
\details{
Relative humidity \eqn{RH} (\%) quantifies the air’s proximity to saturation.
It is calculated using the formula:

\deqn{RH = \frac{e}{e_s} \cdot 100 \approx \frac{w}{w_s} \cdot 100 \approx \frac{q}{q_s} \cdot 100 \approx \frac{\rho_v}{\rho_{vs}} \cdot 100}

where \eqn{e, w, q, \rho_v} is the actual value and
\eqn{e_s, w_s, q_s, \rho_{vs}} is the corresponding saturation value of
vapor pressure, mixing ratio, specific humidity, or absolute humidity.
}
\examples{

# Using vapor pressure (in hPa)
cal_relative_humidity(e = 10, es = 23.4)

# Using specific humidity (in g/kg)
cal_relative_humidity(e = 7, es = 14)

}
\references{
Davis, Robert E., Glenn R. McGregor, and Kyle B. Enfield.
“Humidity: A Review and Primer on Atmospheric Moisture and Human Health.”
Environmental Research 144 (January 2016): 106–16.
\url{https://doi.org/10.1016/j.envres.2015.10.014}.
}
\seealso{
\code{\link[=cal_saturation_vapor_pressure]{cal_saturation_vapor_pressure()}}, \code{\link[=cal_mixing_ratio]{cal_mixing_ratio()}},
\code{\link[=cal_specific_humidity]{cal_specific_humidity()}}, \code{\link[=cal_absolute_humidity]{cal_absolute_humidity()}},
\code{\link[weathermetrics:dewpoint.to.humidity]{weathermetrics::dewpoint.to.humidity()}}
}
