% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_saturation_vapor_pressure.R
\name{cal_saturation_vapor_pressure}
\alias{cal_saturation_vapor_pressure}
\title{Calculate saturation vapor pressure}
\usage{
cal_saturation_vapor_pressure(t, t_unit)
}
\arguments{
\item{t}{Numeric. Air temperature in degree Celsius (°C),
degree Fahrenheit (°F), or Kelvin (K).}

\item{t_unit}{Character string. Unit of the temperature value.
Must be one of \code{"celsius"}, \code{"fahrenheit"}, or \code{"kelvin"}.}
}
\value{
Numeric value of saturation vapor pressure in hPa.
}
\description{
Calculate saturation vapor pressure \eqn{e_s} (hPa) from air temperature
\eqn{T} (°C).
}
\details{
Saturation vapor pressure \eqn{e_s} (hPa) is the vapor pressure of saturated air.
It is calculated using the formula:

\deqn{e_s = 6.108 \cdot exp(\frac{17.27 \cdot T}{T + 237.3})}

where \eqn{T} is air temperature (°C).

The temperature input is converted to degree Celsius (°C) based on
the provided \code{t_unit} before calculation.
}
\examples{
cal_saturation_vapor_pressure(t = 20, t_unit = 'celsius')
cal_saturation_vapor_pressure(t = 68, t_unit = 'fahrenheit')

}
\references{
Davis, Robert E., Glenn R. McGregor, and Kyle B. Enfield.
“Humidity: A Review and Primer on Atmospheric Moisture and Human Health.”
Environmental Research 144 (January 2016): 106–16.
\url{https://doi.org/10.1016/j.envres.2015.10.014}.
}
