% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_temperature_humidity_index.R
\name{cal_temperature_humidity_index}
\alias{cal_temperature_humidity_index}
\title{Calculate temperature-humidity index}
\usage{
cal_temperature_humidity_index(t, tw, t_unit)
}
\arguments{
\item{t, tw}{Numeric. Air temperature and wet-bulb temperature in degree Celsius (°C),
degree Fahrenheit (°F), or Kelvin (K).}

\item{t_unit}{Character string. Unit of the temperature values.
Must be one of \code{"celsius"}, \code{"fahrenheit"}, or \code{"kelvin"}.}
}
\value{
Numeric value of temperature-humidity index in degree Fahrenheit (°F).
}
\description{
Calculate temperature-humidity index \eqn{THI} (°F) from air temperature
\eqn{T} (°F) and wet-bulb temperature \eqn{T_w} (°F).
}
\details{
Temperature-humidity index \eqn{THI} (°F), or “discomfort index”, is an
index used to assess human comfort.
It is calculated using the formula:

\deqn{THI = 0.4 \cdot (T + T_w) + 15}

where \eqn{T} is the air temperature and \eqn{T_w} is the wet-bulb temperature,
in degrees Fahrenheit (°F).

The temperature input is converted to degree Fahrenheit (°F) based on
the provided \code{t_unit} before calculation.
}
\examples{
cal_temperature_humidity_index(t = 68, tw = 52.34, t_unit = 'fahrenheit')
cal_temperature_humidity_index(t = 20, tw = 11.3, t_unit = 'celsius')

}
\references{
Davis, Robert E., Glenn R. McGregor, and Kyle B. Enfield.
“Humidity: A Review and Primer on Atmospheric Moisture and Human Health.”
Environmental Research 144 (January 2016): 106–16.
\url{https://doi.org/10.1016/j.envres.2015.10.014}.

Thom, E. C. “The Discomfort Index.” Weatherwise 12, no. 2 (April 1959): 57–61.
\url{https://doi.org/10.1080/00431672.1959.9926960}.
}
\seealso{
\code{\link[psychrolib:GetTWetBulbFromHumRatio]{psychrolib::GetTWetBulbFromHumRatio()}},
\code{\link[psychrolib:GetTWetBulbFromRelHum]{psychrolib::GetTWetBulbFromRelHum()}},
\code{\link[psychrolib:GetTWetBulbFromTDewPoint]{psychrolib::GetTWetBulbFromTDewPoint()}}
}
