% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_wet_bulb_temperature.R
\name{cal_wet_bulb_temperature}
\alias{cal_wet_bulb_temperature}
\title{Calculate wet-bulb temperature}
\usage{
cal_wet_bulb_temperature(t, td, t_unit)
}
\arguments{
\item{t, td}{Numeric. Air temperature and dewpoint temperature in degree Celsius (°C),
degree Fahrenheit (°F), or Kelvin (K).}

\item{t_unit}{Character string. Unit of the temperature values.
Must be one of \code{"celsius"}, \code{"fahrenheit"}, or \code{"kelvin"}.}
}
\value{
Numeric value of wet-bulb temperature in degree Celsius (°C).
}
\description{
Calculate wet-bulb temperature \eqn{T_w} (°C) from air temperature
\eqn{T} (°C) and dewpoint temperature \eqn{T_d} (°C).
}
\details{
Wet-bulb temperature \eqn{T_w} (°C) is the temperature at which air becomes
saturated by evaporation at constant pressure.
It is calculated using the formula:

\deqn{T_w = T_d + 1/3 * (T - T_d)}

where \eqn{T} is the air temperature and \eqn{T_d} is the dewpoint temperature,
in degrees Celsius (°C).

The temperature input is converted to degree Celsius (°C) based on
the provided \code{t_unit} before calculation.
}
\examples{
cal_wet_bulb_temperature(t = 28, td = 12, t_unit = "celsius")
cal_wet_bulb_temperature(t = 82.4, td = 53.6, t_unit = "fahrenheit")

}
\references{
Davis, Robert E., Glenn R. McGregor, and Kyle B. Enfield.
“Humidity: A Review and Primer on Atmospheric Moisture and Human Health.”
Environmental Research 144 (January 2016): 106–16.
\url{https://doi.org/10.1016/j.envres.2015.10.014}.

Normand, Charles.
“Energy in the Atmosphere.”
Quarterly Journal of the Royal Meteorological Society 72, no. 312–313 (April 1946): 145–67.
\url{https://doi.org/10.1002/qj.49707231202}.
}
