% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unique_value.R
\name{check_unique_value}
\alias{check_unique_value}
\alias{check_unique_value_overview}
\title{Check unique values in a column}
\usage{
check_unique_value(data, column_to_check)

check_unique_value_overview(data, column_to_check)
}
\arguments{
\item{data}{A data frame (for \code{check_unique_value()}) or a list of outputs
from \code{check_unique_value()} (for \code{check_unique_value_overview()}).}

\item{column_to_check}{A column to check for unique values,
using \code{column_name} without quotes.}
}
\value{
\itemize{
\item \code{check_unique_value()}: A tibble with the unique values in
the specified column and their counts.
}

\itemize{
\item \code{check_unique_value_overview()}: A tibble summarizing the number of
unique values in each element of the list and indicating whether
the unique values are consistent across all list elements.
}
}
\description{
\code{check_unique_value()} counts the unique values of a specified column.

\code{check_unique_value_overview()} takes a list of outputs from \code{check_unique_value()}
and checks the consistency of unique values within a specified column across multiple datasets.
}
\examples{

list_mtcars <- list(data1 = mtcars, data2 = mtcars, data3 = mtcars)

res_list <- list_mtcars \%>\%
        purrr::map(check_unique_value, column_to_check = cyl)

check_unique_value_overview(res_list, column_to_check = cyl)

}
