% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts_by_category.R
\name{get_counts_by_category}
\alias{get_counts_by_category}
\title{Get counts by category in wide format}
\usage{
get_counts_by_category(data, group_column, count_column, names_prefix)
}
\arguments{
\item{data}{A data frame.}

\item{group_column}{A column specifying the grouping variable.}

\item{count_column}{A column containing the categories to count within each group.}

\item{names_prefix}{A character string added as a prefix to each of the
resulting column names.}
}
\value{
A tibble in wide format with counts of each category per group.
}
\description{
Summarize the number of occurrences of each category (in \code{count_column})
within each group (defined by \code{group_column}). The result is returned
in a wide format, where each unique category becomes a column.
Missing combinations are filled with 0.
}
\examples{

df <- data.frame(
  Incurred_Date = as.Date(c('2021-01-01', '2021-01-01', '2021-01-02')),
  Status = c("Open", "Closed", "Open")
)
get_counts_by_category(df,
                       group_column = Incurred_Date,
                       count_column = Status,
                       names_prefix = "Status_")

}
