% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Get full working datasets}
\usage{
get_dataset(folder_path, pattern)
}
\arguments{
\item{folder_path}{A string path to the directory containing sub-directories
of input Parquet files. The path should not end with a forward slash (\code{/}).}

\item{pattern}{A string to filter the sub-directories of input files by names.}
}
\value{
A named list of datasets.
}
\description{
Get full working datasets as a named list of
\href{https://arrow.apache.org/docs/r/reference/Dataset.html}{Dataset} R6 objects.
Each dataset is opened from files matching the given pattern, and the names
of the list elements correspond to the file names.
}
\examples{
\dontrun{
datasets <- get_dataset(folder_path = "path/to/data_folder", pattern = "Facility")
}

}
