% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parquet_arrow.R
\name{get_parquet_arrow}
\alias{get_parquet_arrow}
\title{Convert multiple files to Parquet format}
\usage{
get_parquet_arrow(
  folder_path,
  pattern,
  input_format = "csv",
  max_rows_per_file = 1e+07
)
}
\arguments{
\item{folder_path}{A string path to the directory containing
sub-directories of input files.}

\item{pattern}{A string to filter the sub-directories of input files by names.}

\item{input_format}{A string indicating format of input files.
Default is \code{"csv"}. Other supported formats are listed in \code{\link[arrow:open_dataset]{arrow::open_dataset()}}.}

\item{max_rows_per_file}{Maximum number of rows per output Parquet file.
Default is \code{1e7} (10 million rows).}
}
\value{
Converted Parquet files, saved in original location of input files.
}
\description{
Convert multiple files from a specified format (e.g., CSV) to Parquet format.
The converted Parquet files are saved in the original location of the input files,
allowing for efficient storage and retrieval of large datasets.
}
\examples{
\dontrun{

# Assuming 'folderX' containing 10 sub-folders
# (3 folders of city A, 3 folders of city B, and 4 folders of city C) with CSV files,
# convert all the data of city C to Parquet format:

get_parquet_arrow(folder_path = "path/to/folderX",
                  pattern = "city_C")
}

}
\seealso{
\code{\link[arrow:open_dataset]{arrow::open_dataset()}}, \code{\link[arrow:write_dataset]{arrow::write_dataset()}}
}
