% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_era5_temperature}
\alias{load_era5_temperature}
\title{Load hourly ERA5-Land temperature data for the Richmond MSA}
\usage{
load_era5_temperature()
}
\value{
A \code{SpatRaster}
}
\description{
Load a \code{SpatRaster} object containing hourly temperature data
from \href{https://cds.climate.copernicus.eu/datasets/reanalysis-era5-land?tab=overview}{ERA5-Land}
for the Richmond Metropolitan Statistical Area (MSA).
The time dimension in the raster spans from 2017-05-14 00:00:00 to
2017-05-16 23:00:00 Eastern Time.
}
\examples{
r <- load_era5_temperature()
}
