% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_values.R
\name{recode_values}
\alias{recode_values}
\title{Recode values based on grouping logic}
\usage{
recode_values(data, var1, var2, unknown_category)
}
\arguments{
\item{data}{A data frame or an element of a list dataset
(output from \code{\link[=get_dataset]{get_dataset()}}).}

\item{var1}{A character string specifying the grouping variable.}

\item{var2}{A character string specifying the variable to be recoded.}

\item{unknown_category}{A character string specifying the ambiguous or
unknown values in \code{var2}.}
}
\value{
A data frame with \code{var1} and recoded \code{var2}.
}
\description{
This function groups data by \code{var1} and recodes \code{var2} based on specific conditions.
If \code{var1} has only one unique \code{var2} value, the original value is retained.
If \code{var1} has two unique \code{var2} values, the known value is assigned if one matches \code{unknown_category};
otherwise, \code{unknown_category} is assigned.
If \code{var1} has more than two unique \code{var2} values, \code{unknown_category} is assigned.
}
\examples{
df <- data.frame(ID = c(1, 1, 2, 2, 3, 3, 3),
                 Status = c("Male", "Unknown", "Female", "Female", "Male", "Female", "Unknown"))
df

recode_values(df, var1 = "ID", var2 = "Status", unknown_category = "Unknown")

}
