% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_va_zcta.R
\docType{data}
\name{sf_va_zcta}
\alias{sf_va_zcta}
\title{Geographic boundaries of ZCTAs in Virginia}
\format{
\subsection{\code{sf_va_zcta}}{

\describe{
\item{ZCTA5CE20}{Five-digit ZCTA code (2020).}
\item{zcta_county}{County or independent city associated with the ZCTA.}
\item{geometry}{Polygon geometry of the ZCTA.}
}
}
}
\source{
Data originally obtained from the \href{https://www.census.gov/programs-surveys/geography/guidance/geo-areas/zctas.html}{U.S. Census Bureau, 2020 TIGER/Line Shapefiles},
and post-processed to get the ZCTAs in Virginia.
}
\usage{
sf_va_zcta
}
\description{
A simple feature collection of ZIP Code Tabulation Areas (ZCTAs) in Virginia.
}
\keyword{datasets}
