% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_county_health_regions.R
\docType{data}
\name{va_county_health_regions}
\alias{va_county_health_regions}
\title{Virginia county and Health Planning Region crosswalk}
\format{
\subsection{\code{va_county_health_regions}}{

\describe{
\item{County}{Short name of the county or independent city.}
\item{County_full_name}{Full official name of the county or independent city.}
\item{Health_Planning_District}{Virginia Health Planning District.}
\item{Health_Planning_Region}{Virginia Health Planning Region.}
\item{MSA_Rating_Area}{Metropolitan Statistical Area (Rating Area).}
}
}
}
\source{
\href{https://www.scc.virginia.gov/media/sccvirginiagov-home/applications-assets/aca-rating-area-map.pdf}{Virginia Rating Areas map}
provided by the Virginia State Corporation Commission (SCC).

\href{https://www.vdh.virginia.gov/local-health-districts/}{Local Health Districts}
provided by Virginia Department of Health.
}
\usage{
va_county_health_regions
}
\description{
A reference dataset of 133 counties and independent cities in Virginia and
their associated health planning geographies (Health Planning District,
Health Planning Region, and Metropolitan Statistical Area (MSA) Rating Area).
}
\keyword{datasets}
