% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_obs.R
\name{get_weather_obs}
\alias{get_weather_obs}
\title{Get observation closest to a specific time}
\usage{
get_weather_obs(
  data,
  target_column,
  timestamp_column,
  target_hour,
  target_minute,
  time_window
)
}
\arguments{
\item{data}{A data frame containing weather observations.}

\item{target_column}{The meteorological variable to retrieve.}

\item{timestamp_column}{The column containing timestamps of the observations.}

\item{target_hour}{An integer specifying the target hour (0-23)
for retrieving observations.}

\item{target_minute}{An integer specifying the target minute (0-59)
for retrieving observations.}

\item{time_window}{An integer specifying the allowable time window
(in minutes) around the target time.}
}
\value{
A numeric value or \code{NA}.
}
\description{
Extract the maximum available weather observation closest to a
specified target time within a given time window.
If no valid observations are found, return \code{NA}.
}
\examples{
df <- data.frame(
  timestamp = (c("2024-03-01 06:50:00",
                 "2024-03-01 07:10:00",
                 "2024-03-01 07:30:00",
                 "2024-03-01 07:55:00")),
  temperature = c(15, 16, NA, 17)
)

# retrieve the closest temperature observation around 07:00 within a 30-minute window
get_weather_obs(df,
                target_column = temperature,
                timestamp_column = timestamp,
                target_hour = 7,
                target_minute = 0,
                time_window = 30)

}
\references{
Davis, R.E., Himmel, O., Sims, P.K. et al. Mortality and diurnal temperature
range in Virginia. Int J Biometeorol (2025).
\url{https://doi.org/10.1007/s00484-025-02850-6}
}
